Mime-new screen has two use cases: 1. Loading a previously created mime 2. Creating a new mime
When you go to the new screen and a mimeID is sent, then it is seen as previously created. If a mimeId is not sent in the state, then it means to create a new mime.

I also decoupled the mime generation, and updated the UI a little more to be showcasing the status.